updated database

+++ Geekium/Models/AccountPurchase.cs
@@ -14,7 +14,7 @@ namespace Geekium.Models
         public int SellerId { get; set; }
         public int CartId { get; set; }
         public DateTime PurchaseDate { get; set; }
-        public int PurchasePrice { get; set; }
+        public double PurchasePrice { get; set; }
         public int TrackingNumber { get; set; }
         public string SellerName { get; set; }
         public int PointsGained { get; set; }
