+++ Geekium/Models/SellerReview.cs
@@ -11,7 +11,7 @@ namespace Geekium.Models
     {
         public int SellerReviewId { get; set; }
         public int SellerId { get; set; }
-        public int? BuyerRating { get; set; }
+        public double? BuyerRating { get; set; }
         public string ReviewDescription { get; set; }
 
         public virtual SellerAccount Seller { get; set; }
